test: introduce TemplateResolver mock
