+++ modules/angular2/test/directives/foreach_spec.js
@@ -1,8 +1,7 @@
 import {describe, xit, it, expect, beforeEach, ddescribe, iit, el} from 'angular2/test_lib';
 
 import {DOM} from 'angular2/src/facade/dom';
-import {Map, MapWrapper, ListWrapper} from 'angular2/src/facade/collection';
-import {Type, isPresent} from 'angular2/src/facade/lang';
+import {ListWrapper} from 'angular2/src/facade/collection';
 
 import {Injector} from 'angular2/di';
 import {Lexer, Parser, ChangeDetector, dynamicChangeDetection} from 'angular2/change_detection';
@@ -11,10 +10,11 @@ import {Compiler, CompilerCache} from 'angular2/src/core/compiler/compiler';
 import {DirectiveMetadataReader} from 'angular2/src/core/compiler/directive_metadata_reader';
 import {NativeShadowDomStrategy} from 'angular2/src/core/compiler/shadow_dom_strategy';
 import {TemplateLoader} from 'angular2/src/core/compiler/template_loader';
-import {TemplateResolver} from 'angular2/src/core/compiler/template_resolver';
 
-import {Decorator, Component, Viewport} from 'angular2/src/core/annotations/annotations';
 import {Template} from 'angular2/src/core/annotations/template';
+import {Decorator, Component, Viewport} from 'angular2/src/core/annotations/annotations';
+
+import {MockTemplateResolver} from 'angular2/src/mock/template_resolver_mock';
 
 import {Foreach} from 'angular2/src/directives/foreach';
 
@@ -22,7 +22,7 @@ export function main() {
   describe('foreach', () => {
     var view, cd, compiler, component, tplResolver;
     beforeEach(() => {
-      tplResolver = new FakeTemplateResolver();
+      tplResolver = new MockTemplateResolver();
       compiler = new Compiler(dynamicChangeDetection, new TemplateLoader(null),
         new DirectiveMetadataReader(), new Parser(new Lexer()), new CompilerCache(),
         new NativeShadowDomStrategy(), tplResolver);
@@ -232,28 +232,5 @@ class TestComponent {
   item: any;
   constructor() {
     this.items = [1, 2];
-  }
-}
-
-class FakeTemplateResolver extends TemplateResolver {
-  _cmpTemplates: Map;
-
-  constructor() {
-    super();
-    this._cmpTemplates = MapWrapper.create();
-  }
-
-  setTemplate(component: Type, template: Template) {
-    MapWrapper.set(this._cmpTemplates, component, template);
-  }
-
-  resolve(component: Type): Template {
-    var override = MapWrapper.get(this._cmpTemplates, component);
-
-    if (isPresent(override)) {
-      return override;
-    }
-
-    return super.resolve(component);
   }
 }
